Using float & finding average of point values

+++ 153204buggycode.cpp
@@ -6,17 +6,17 @@ using namespace std;
 int main()
 {
 
-	int nSum=0;
-	int nNums;
+	float nSum=0.0;
+	float nNums;
 
 	// accumulate input numbers until the
 	// user enteres a negative number, then
 	// return the average
-	nNums = 0;
+	nNums = 0.0;
 	while(true)
 	{
 		// enter another number to add
-		int nValue;
+		float nValue;
 		cout << "Enter another number:";
 		cin >> nValue;
 		cout << endl;
